modules/spokes: fix missing spoke in some counts
Due to floating point errors this loop would occasionally miss
the last spoke as-is.

A quick fix is to just round down the accumulator before
comparing it... this function in general needs to all be cleaned
up, esp. since the conversion to floats.

An obvious repro of this issue, fixed by this commit:

'--module=spokes,iterations=2,count=3,twist=0.125,thickness=2' \
'--video=sdl,ratio=full,fullscreen=off,vsync=on,size=640x480'

The low count of 3 makes the missing spoke quite visible...