drm_fb: suppress drmAvailable() checks
This is preventing --video=drm from working on my X230 today for
no good reason.  It seems libdrm is assuming /dev/dri/card0 will
be present if drm's available.  But for some reason my X230 only
has /dev/dri/card1 presently.  It didn't used to be this way,
something is weird, but I'm not going to dig into it.  If it
happened here, it may happen elsewhere, and drmAvailable() is
proving actively harmful.

For now I'm just disabling the check w/#if 0