modules/flow: fix edge case segfault
When flow became threaded this bug was introduced.  It's possible
for the unchecked put_pixel to go out of bounds while stepping
through n_iters, since the bounds check bypassing this doesn't
consider the full range of the iterated coordinates.

A simple fix to prevent the segfault is to use the checked
variant.